tenant/api/requirements.txt
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.4
psycopg2==2.7.3
pytz==2017.2
